omni-enterprise-ultra-max/backend/requirements.txt
# Core Framework
fastapi==0.115.4
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-dotenv==1.0.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.36
alembic==1.13.0
psycopg2-binary==2.9.9
pymysql==1.1.0
pymongo>=4.0,<5.0
motor==3.3.2
redis[asyncio]==5.2.0

# Authentication & Security
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pyotp==2.9.0
cryptography==43.0.3

# Payment Gateways
stripe==11.1.1

# Google Cloud Platform
google-cloud-firestore==2.13.1
google-cloud-storage==2.13.0
google-cloud-secret-manager==2.17.0
google-generativeai==0.8.3

# Communications
sendgrid==6.11.0
twilio==8.10.0
python-telegram-bot==20.7

# Utilities
requests==2.31.0
# Align with python-telegram-bot 20.7 which requires httpx~=0.25.2
httpx==0.27.2
email-validator>=2.0.0
pillow==10.1.0
python-magic==0.4.27

# AI/ML Core
openai==1.54.4
anthropic==0.39.0
tensorflow==2.17.1
torch==2.5.1
torchvision==0.20.1

# ML & Data Science
scikit-learn==1.5.2
pandas==2.2.3
numpy==1.26.2
scipy==1.11.4

# Time Series & Forecasting
prophet==1.1.5
statsmodels==0.14.0

# NLP & Language Models
spacy==3.7.2
transformers==4.46.3
sentence-transformers==3.3.1
nltk==3.8.1

# Vector Search & Embeddings
faiss-cpu==1.7.4
chromadb==0.4.18

# Anomaly Detection
pyod==1.1.2

# Graph Database (Neo4j)
neo4j==5.14.1
py2neo==2021.2.4

# Deep Learning Utilities
keras==2.15.0
xgboost==2.0.2
lightgbm==4.1.0

# Computer Vision
opencv-python==4.8.1.78

# Monitoring & Logging
sentry-sdk==2.18.0
prometheus-client==0.21.0

# Task Queue
celery==5.3.4
kombu==5.3.4

# Distributed Tracing (OpenTelemetry)
opentelemetry-api==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-exporter-jaeger==1.21.0

# Test dependencies
pytest>=7.0
pytest-mock>=3.10.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
requests-mock>=1.9.0
faker==20.1.0